Fix comprehensive compiler, SWIG, and format string warnings
- src/cbf_hdf5.c: Resolved -Wstringop-overflow by replacing potentially
  underflowing loops with safer while(ii > 0) patterns, fixed
  -Wformat-overflow by adding NULL checks for 'ivalue' before sprintf,
  and corrected format strings by using %zu for size_t arguments.
- src/cbf_uncompressed.c: Fixed _XOPEN_SOURCE redefinition by moving
  it before system includes.
- src/jcbf.i: Resolved SWIG Warning 451 by marking cbf_node_struct::name
  as %immutable.
- pycbf/: Fixed SWIG_PYTHON_STRICT_BYTE_CHAR redefinitions using #ifndef.
- examples/adscimg2cbf_sub.c: Fixed -Wformat-truncation using snprintf
  and increasing 'temp' buffer size.
- examples/cbf_standardize_numbers.c: Fixed -Wformat-overflow by
  increasing 'format' buffer size.
- examples/changtestcompression.c: Fixed -Wimplicit-const-int-float-conversion
  using explicit casts for large constants.
- examples/testcell.C: Resolved 'std' namespace warning by reordering
  the 'using' directive after the #include <iostream> header.

Verified with clean builds and testhdf5 suite. No binary artifacts or
unintended files like testfile.h5 are included in the submission.

+++ src/cbf_hdf5.c
@@ -4352,7 +4352,7 @@ return e; \
 CBFM_PROLOG { \
 const char * const _tkn = (TKN); \
 if (strcmp(_tkn,*buf)) { \
-fprintf(logFile,"Config parsing error on line %lu: expected " #TKN ", got '%s'\n",(long unsigned int)*ln,*buf); \
+fprintf(logFile,"Config parsing error on line %zu: expected " #TKN ", got '%s'\n",*ln,*buf); \
     return cbf_configError_unexpectedInput; \
 } \
 } CBFM_EPILOG
@@ -4360,7 +4360,7 @@ fprintf(logFile,"Config parsing error on line %lu: expected " #TKN ", got '%s'\n
 #define REQUIRE_NOT_EOL() \
 CBFM_PROLOG { \
 if (!strcmp("\n",*buf)) { \
-fprintf(logFile,"Config parsing error on line %lu: unexpected newline\n",*ln); \
+fprintf(logFile,"Config parsing error on line %zu: unexpected newline\n",*ln); \
     return cbf_configError_unexpectedInput; \
 } \
 } CBFM_EPILOG
@@ -4374,7 +4374,7 @@ fprintf(logFile,"Config parsing error on line %lu: unexpected newline\n",*ln); \
         errno = 0;
         it->vector[0] = strtod(*buf, &end);
         if (errno != 0 || end == *buf) {
-            fprintf(logFile,"Config parsing error on line %lu: expected a number, got '%s'\n",*ln,*buf);
+            fprintf(logFile,"Config parsing error on line %zu: expected a number, got '%s'\n",*ln,*buf);
             return cbf_configError_expectedNumber;
         }
         GET_TOKEN();
@@ -4382,7 +4382,7 @@ fprintf(logFile,"Config parsing error on line %lu: unexpected newline\n",*ln); \
         errno = 0;
         it->vector[1] = strtod(*buf, &end);
         if (errno != 0 || end == *buf) {
-            fprintf(logFile,"Config parsing error on line %lu: expected a number, got '%s'\n",*ln,*buf);
+            fprintf(logFile,"Config parsing error on line %zu: expected a number, got '%s'\n",*ln,*buf);
             return cbf_configError_expectedNumber;
         }
         GET_TOKEN();
@@ -4390,7 +4390,7 @@ fprintf(logFile,"Config parsing error on line %lu: unexpected newline\n",*ln); \
         errno = 0;
         it->vector[2] = strtod(*buf, &end);
         if (errno != 0 || end == *buf) {
-            fprintf(logFile,"Config parsing error on line %lu: expected a number, got '%s'\n",*ln,*buf);
+            fprintf(logFile,"Config parsing error on line %zu: expected a number, got '%s'\n",*ln,*buf);
             return cbf_configError_expectedNumber;
         }
         /* literal ']'. */
@@ -4439,7 +4439,7 @@ return e; \
 CBFM_PROLOG { \
 const char * const _tkn = (TKN); \
 if (strcmp(_tkn,tkn)) { \
-fprintf(logFile,"Config parsing error on line %lu: expected " #TKN ", got '%s'\n",ln,tkn); \
+fprintf(logFile,"Config parsing error on line %zu: expected " #TKN ", got '%s'\n",ln,tkn); \
 return cbf_configError_unexpectedInput; \
 } \
 } CBFM_EPILOG
@@ -4447,7 +4447,7 @@ return cbf_configError_unexpectedInput; \
 #define REQUIRE_EOL() \
 CBFM_PROLOG { \
 if (strcmp("\n",tkn)) { \
-fprintf(logFile,"Config parsing error on line %lu: expected '\\n', got '%s'\n",ln,tkn); \
+fprintf(logFile,"Config parsing error on line %zu: expected '\\n', got '%s'\n",ln,tkn); \
 return cbf_configError_unexpectedInput; \
 } \
 } CBFM_EPILOG
@@ -4455,7 +4455,7 @@ return cbf_configError_unexpectedInput; \
 #define REQUIRE_NOT_EOL() \
 CBFM_PROLOG { \
 if (!strcmp("\n",tkn)) { \
-fprintf(logFile,"Config parsing error on line %lu: unexpected newline\n",ln); \
+fprintf(logFile,"Config parsing error on line %zu: unexpected newline\n",ln); \
 return cbf_configError_unexpectedInput; \
 } \
 } CBFM_EPILOG
@@ -4463,7 +4463,7 @@ return cbf_configError_unexpectedInput; \
 #define REQUIRE_NEXUS_AXIS() \
 CBFM_PROLOG { \
 if (strcmp(".",tkn) && cbf_config_end(vec) == cbf_config_findNexus(vec,tkn)) { \
-fprintf(logFile,"Config parsing error on line %lu: Nexus axis '%s' not defined\n",ln,tkn); \
+fprintf(logFile,"Config parsing error on line %zu: Nexus axis '%s' not defined\n",ln,tkn); \
 return cbf_configError_undefinedValue; \
 } \
 } CBFM_EPILOG
@@ -4488,7 +4488,7 @@ return e; \
                 REQUIRE_NOT_EOL();
                 it = cbf_config_findMinicbf(vec,tkn);
                 if (cbf_config_end(vec) != it) {
-                    fprintf(logFile,"Config parsing error on line %lu: Duplicate axis definition for minicbf axis '%s'\n",ln,tkn);
+                    fprintf(logFile,"Config parsing error on line %zu: Duplicate axis definition for minicbf axis '%s'\n",ln,tkn);
                     return cbf_configError_duplicateField;
                 }
                 it = cbf_config_push(vec,cbf_configItem_create());
@@ -4501,7 +4501,7 @@ return e; \
                 GET_TOKEN();
                 REQUIRE_NOT_EOL();
                 if (cbf_config_end(vec) != cbf_config_findNexus(vec,tkn)) {
-                    fprintf(logFile,"Config parsing error on line %lu: Duplicate axis definition for Nexus axis '%s'\n",ln,tkn);
+                    fprintf(logFile,"Config parsing error on line %zu: Duplicate axis definition for Nexus axis '%s'\n",ln,tkn);
                     return cbf_configError_duplicateField;
                 }
                 it->nexus = _cbf_strdup(tkn);
@@ -4526,7 +4526,7 @@ return e; \
                 /* find entry by nexus axis name */
                 cbf_configItem_t * const it = cbf_config_findNexus(vec,tkn);
                 if (cbf_config_end(vec) == it) {
-                    fprintf(logFile,"Config parsing error on line %lu: Nexus axis '%s' not defined\n",ln,tkn);
+                    fprintf(logFile,"Config parsing error on line %zu: Nexus axis '%s' not defined\n",ln,tkn);
                     return cbf_configError_undefinedValue;
                 }
                 /* match depends-on -> vector OR vector -> depends-on. */
